+++ SyncService/PluginValidator/IPluginValidator.cs
@@ -1,11 +1,10 @@
 ﻿using XrmSync.Model.Plugin;
+using XrmSync.Model.CustomApi;
 
 namespace XrmSync.SyncService.PluginValidator;
 
 public interface IPluginValidator
 {
     void Validate(List<PluginDefinition> pluginTypes);
-
-    // TODO: Implement validation for custom APIs
-    //void Validate(List<ApiDefinition> customApis);
+    void Validate(List<CustomApiDefinition> customApis);
 }