+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Validation occurs before save-config and before main operations
   - Clear error messages with detailed validation feedback
 * Refactoring of the internal model to better reflect the actual data and not drag around a large number of loosely coupled lists.
+* Validation of CustomAPIs in addition to Plugins
 
 ### v1.0.0-preview.1 - 15 August 2025
 * Initial version